Use ints in BitInputStream instead of bytes they are too awkward

+++ src/main/java/ie/davidloftus/huffman/BitInputStream.java
@@ -24,7 +24,7 @@ public class BitInputStream {
         return byteBuffer;
     }
 
-    public byte readBit() throws IOException {
+    public int readBit() throws IOException {
         if (currentBit == -1) {
             nextByte();
         }
